09.19 18:25 최정준

+++ dasony/src/main/java/com/ds/dasony/game/model/vo/Game.java
@@ -15,6 +15,6 @@ public class Game {
 	private String gameStatus;
 	private String pointStatus;
 	private String ticketStatus;
-	private long userNo;
+	private int userNo;
 	
 }
